fix: 修复四个bug
【欧拉版】【ARM】解析重命名-资源解析窗口可对不同任务重复命名，而且重复新建多个任务，原有已被修改名称的任务名又恢复至初始状态 【欧拉版】
【ARM】【默认下载路径】下载过程中修改文件夹名称 【欧拉版】【ARM】【解析框】下载错误图片链接可以解析下载失败 【欧拉版】【ARM】【设置】
下载通知点击查看没有反应

Log：【欧拉版】【ARM】解析重命名-资源解析窗口可对不同任务重复命名，而且重复新建多个任务，原有已被修改名称的任务名又恢复至初始状态 【欧拉
版】【ARM】【默认下载路径】下载过程中修改文件夹名称 【欧拉版】【ARM】【解析框】下载错误图片链接可以解析下载失败 【欧拉版】【ARM】【设
置】下载通知点击查看没有反应

+++ src/src/ui/createTask/urlthread.cpp
@@ -169,7 +169,7 @@ void UrlThread::onHttpRequest(QNetworkReply *reply)
     {
         QProcess process;
         QStringList list;
-        list << "-i" << reply->url().toString();
+        list << "-i" << QString("\"" +reply->url().toString() + "\"");
         process.start("curl", list);
         connect(&process, &QProcess::readyReadStandardOutput, this, [=]() {
             static QMutex mutex;
@@ -206,16 +206,24 @@ void UrlThread::onHttpRequest(QNetworkReply *reply)
         list << "-I" << "-k" << reply->url().toString();
         QString str = reply->url().toString();
         process->start("curl", list);
+        qDebug()<< list;
         connect(process, &QProcess::readyReadStandardOutput, this, [=]() {
             qDebug()<<"readyReadStandardOutput";
             static QMutex mutex;
             mutex.lock();
             QString str =process->readAllStandardOutput();
+            QStringList list = str.split("\r\n");
+            if(!list[0].contains("200")){
+                mutex.unlock();
+                return ;
+            }
             process->kill();
             process->close();
             delete process;
             //process = nullptr;
+
             m_linkInfo.urlSize = getUrlSize(str);
+            m_linkInfo.type =getUrlType(str);
             emit sendFinishedUrl(m_linkInfo);
             mutex.unlock();
